% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_regression.R
\name{gene_regression}
\alias{gene_regression}
\title{Construct data for gene regression}
\usage{
gene_regression(
  count,
  indmatrix,
  selected = NULL,
  neg,
  gene_frac = 0.01,
  slot = "scale.data"
)
}
\arguments{
\item{count}{A Seurat object.Noted that the object has to be normalized and scaled by standard procedure.}

\item{indmatrix}{A perturbation-cell matrix calculated by indmatrix function.}

\item{selected}{Interested features list. Default: NULL.}

\item{neg}{A parameter for control cell. Non-perturbed cells are recommended to assign NT.}

\item{gene_frac}{A parameter for filtering low expression genes. Default: 0.01.}

\item{slot}{A parameter for extracting data from Seurat object. Default: 'scale.data'.}
}
\value{
A list for gene regression. Xmat represents cell identity matrix that contains cell and perturbation parameters. Ymat represents gene count matrix.
}
\description{
Construct data for gene regression
}
\examples{
mat_reg <- gene_regression(count, ind, selected = NULL,neg,
                            gene_frac = 0.01, slot = 'scale.data')
}
