% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indmatrix.R
\name{indmatrix}
\alias{indmatrix}
\title{Extract perturbation-cell matrix}
\usage{
indmatrix(count, sgrna)
}
\arguments{
\item{count}{A Seurat object. Noted that the object has to be normalized and scaled by standard procedure.}

\item{sgrna}{A perturbation data frame which contains perturbation, cell parameters.}
}
\value{
A perturbation-cell matrix
}
\description{
Construct perturbation-cell matrix. The cell is assigned with TRUE which is infected by corresponding perturbation.
}
\examples{
ind <- indmatrix(count, sgrna)
}
