% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_permutation.R
\name{matrix_permutation}
\alias{matrix_permutation}
\title{Gene regression}
\usage{
matrix_permutation(X, Y, lamb = 0.01, permut = 1000)
}
\arguments{
\item{X}{A cell identity matrix calculated by gene_regression function.}

\item{Y}{A gene count matrix calculated by gene_regression function.}

\item{lamb}{The Regularization coefficient. Default: 0.01.}

\item{permut}{The number of permutations for p value. Default: 1000.}
}
\value{
A list for regression result.
}
\description{
The effect of perturbation is calculated by gene ridge regression.
}
\examples{
perscore <- matrix_permutation(X, Y, lamb = 0.01, permut = 1000)
}
