% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_score.R
\name{regression_score}
\alias{regression_score}
\title{Calculate perturbation effect}
\usage{
regression_score(
  count,
  sgrna,
  negctr,
  selected_gene = NULL,
  permut = NULL,
  paramt = 0.01,
  filt = 0.01,
  slot = "scale.data"
)
}
\arguments{
\item{count}{A Seurat object. Noted that the object has to be normalized and scaled by standard procedure.}

\item{sgrna}{A perturbation-cell data frame. Noted that the data frame contain perturbation and cell parameters.}

\item{negctr}{A parameter for non-perturbed cells which is recommended as NT.}

\item{selected_gene}{Interested features list. Default: NULL.}

\item{permut}{The number of permutations for p value. Default: NULL.}

\item{paramt}{The Regularization coefficient. Default: 0.01.}

\item{filt}{A parameter for filtering low expression genes. Default: 0.01.}

\item{slot}{A parameter for extracting data from Seurat object. Default: 'scale.data'.}
}
\value{
A list for regression result.
}
\description{
Calculate perturbation effect by ridge regression in single-cell CRISPR screening data.
}
\examples{
reg_score <- regression_score(count,sgrna,negctr,selected_gene = NULL,
                               permut = 1000,paramt = 0.01,filt = 0.01,slot = 'scale.data')
}
